creating new file for count down timer
without user input
